requirements.txt
azure-functions
psycopg[binary]>=3.1,<4
